costs

+++ GroundFunc.cpp
@@ -6,7 +6,7 @@ void GroundFunc<double>::PDDLPrint( std::ostream & s, unsigned indent, const Tok
 	tabindent( s, indent );
 	s << "( = ";
 	TypeGround::PDDLPrint( s, 0, ts, d );
-	s << " " << value << " )";
+	s << " " << ( int )value << " )";
 }
 
 template <>
